moved user settings controllers into one file
