+++ server/routes/auth.js
@@ -2,9 +2,7 @@ const express = require('express');
 const router = express.Router();
 const login = require('../controllers/login');
 const signup = require('../controllers/signup');
-const changeUserName = require('../controllers/changeUserName');
-const changePassword = require('../controllers/changePassword');
-const ChangeAvatar = require('../controllers/changeAvatar');
+const userSettings = require('../controllers/userSettings');
 const { authUser } = require('../controllers/authLogin');
 const verifyToken = require('../middlewares/authentication');
 
@@ -16,11 +14,11 @@ router.post('/signup', (req, res) => {
   signup.handleRegister(req, res);
 });
 
-router.post('/changeusername', verifyToken, changeUserName.handleChangeUserName);
+router.post('/changeusername', verifyToken, userSettings.handleChangeUserName);
 
-router.post('/changepassword', verifyToken, changePassword.handleChangePassword);
+router.post('/changepassword', verifyToken, userSettings.handleChangePassword);
 
-router.post('/changeavatar', verifyToken, ChangeAvatar.handleChangeAvatar);
+router.post('/changeavatar', verifyToken, userSettings.handleChangeAvatar);
 
 router.get('/authLogin', authUser);
